seo-tools/requirements.txt
dify_plugin>=0.2.0
requests>=2.25.0
python-dotenv>=0.19.0